Improve file heirarchy management
Improve recursive file heirarchy management such that the deletion
of an Encounter will delete all it's recordings and the delection
of a Recording will delete all it's files.

Add a trash based deletion system such that any files deleted are
moved to a temporary trash folder with a flat structure. Each file
is unique by it's assigned unique file UUID. Still need to create
code to interact with the trash system.