Add contour route file and selection table cross referencing
Add a contour routes file (currently empty with no routes, other
than initialisation). Add routes_contour.py as a blueprint in
app.py.

Selection tables must be cross referenced with the actual uploaded
selections. The scaffold for this is done in routes_recording. Must
still implement the parsing of the selection table CSV.

Modify the database with new fields in each selection (pertaining to
a contour file). Also add a field for a contour file ID. The database
must be re-initialised using create_database.sql in this version of
the code.