Add version control management for recording table
Using the Maria DB System Versioning system the database recording
table was modified to track each time it is changed. In the user
interface for each recording a table is shown with the log of all
change date-times, actions and users. As of yet this is not role-
restricted but should be to System and Database Administrators.