Add version control to selection and fix selection table
Add a version control table to the selection-view.html page. This
allows the user to enter archive mode based on when changes were
made to a selection. This works much like the functionality already
built for the recording archive.

Change selection table. For the archive mode to work the selection
table was ported into the Maria DB database (whereas previously the
code would store the selection table itself and parse through it when
required). Now the code copies all data from the selection table over
to existing or new Selection objects.