Fix archival system
Fix archival system, including that of Encounter and Recording. N-
ow the user can enter an archival view of the entire program at a-
ny point in history. However if a user attempts to view an archiv-
al view of the program at a point in which there was no archive t-
hey are greeted with a generic error.