Add error page for invalid URL
Whenever the user requests a URL of an item that exists in live
but they are viewing in archive, or the other way around (i.e.
the item does not exist in their view), show them an error message
and allow them to either return to the home page or, if they are
in the archive view, disable it.