Prepare OCEAN for test environment
Create an automatic pytest script in testing folder.

Update app.py and database_handler.py to separate flask and database
functions separately. This modularisation is beneficial for automated
testing which requires a simulation of the web application.

Move all routes that were in app.py to routes_general.py (again -
improving modularisation of code). All links to these routes are
thought to have been refactored, but there may be a few that still
need changing.

Update README with updated setup instructions as well as testing
information.