Add support for single-sign-on
Change the authentication method for OCEAN to support single sign on.
Now authentication takes place in main.py (rather than routes_auth.py)
and uses the single sign on email to authenticate access to the program.
Note that this does not work in the development environment. It is for
this reason that in main.sso_login(), the authentication email can be
overriden.