Make error handling more robust
If the folder for the filespace is not found, OCEAN will now
still start, and remain running. Any components relying on
the filespace will simply produce an error message.

The database is pinged every five minutes to ensure the server
keeps the connection to the client open.

Error messages are displayed as simple text, and with more info.
Previously a database error was causing an infinite loop because
OCEAN tried to load user data to associate with the error message.
Querying the databsae for this user data would cause another error,
and so on...

The logger file is now protected in logger.py. This ensures that
changes are immediately flushed. Previously this was very
inconsistent.